Follow jruby version API of ExtensionValue
* Removed Extended.create
* Still moved the spec to cruby since jruby's ExtensionValue's type and
  payload can't be accessed.
